+++ tapo-py/src/handlers.rs
@@ -1,7 +1,9 @@
 mod generic_device_handler;
+mod light_handler;
 mod plug_energy_monitoring_handler;
 mod plug_handler;
 
 pub use generic_device_handler::*;
+pub use light_handler::*;
 pub use plug_energy_monitoring_handler::*;
 pub use plug_handler::*;