+++ tapo-py/src/handlers/light_handler.rs
@@ -1,29 +1,19 @@
 use std::sync::Arc;
 
-use chrono::NaiveDate;
 use pyo3::prelude::*;
-use tapo::requests::EnergyDataInterval;
-use tapo::PlugEnergyMonitoringHandler;
+use tapo::LightHandler;
 use tokio::sync::Mutex;
 
 use crate::errors::ErrorWrapper;
 
 #[derive(Clone)]
-#[pyclass(name = "EnergyDataInterval")]
-pub enum PyEnergyDataInterval {
-    Hourly,
-    Daily,
-    Monthly,
+#[pyclass(name = "LightHandler")]
+pub struct PyLightHandler {
+    handler: Arc<Mutex<LightHandler>>,
 }
 
-#[derive(Clone)]
-#[pyclass(name = "PlugEnergyMonitoringHandler")]
-pub struct PyPlugEnergyMonitoringHandler {
-    handler: Arc<Mutex<PlugEnergyMonitoringHandler>>,
-}
-
-impl PyPlugEnergyMonitoringHandler {
-    pub fn new(handler: PlugEnergyMonitoringHandler) -> Self {
+impl PyLightHandler {
+    pub fn new(handler: LightHandler) -> Self {
         Self {
             handler: Arc::new(Mutex::new(handler)),
         }
@@ -31,7 +21,7 @@ impl PyPlugEnergyMonitoringHandler {
 }
 
 #[pymethods]
-impl PyPlugEnergyMonitoringHandler {
+impl PyLightHandler {
     pub fn refresh_session<'a>(&'a self, py: Python<'a>) -> PyResult<&'a PyAny> {
         let handler = self.handler.clone();
         pyo3_asyncio::tokio::future_into_py(py, async move {
@@ -102,57 +92,16 @@ impl PyPlugEnergyMonitoringHandler {
         })
     }
 
-    pub fn get_current_power<'a>(&'a self, py: Python<'a>) -> PyResult<&'a PyAny> {
+    pub fn set_brightness<'a>(&'a self, py: Python<'a>, brightness: u8) -> PyResult<&'a PyAny> {
         let handler = self.handler.clone();
         pyo3_asyncio::tokio::future_into_py(py, async move {
-            let result = handler
+            handler
                 .lock()
                 .await
-                .get_current_power()
+                .set_brightness(brightness)
                 .await
                 .map_err(ErrorWrapper)?;
-            Ok(result)
-        })
-    }
-
-    pub fn get_energy_usage<'a>(&'a self, py: Python<'a>) -> PyResult<&'a PyAny> {
-        let handler = self.handler.clone();
-        pyo3_asyncio::tokio::future_into_py(py, async move {
-            let result = handler
-                .lock()
-                .await
-                .get_energy_usage()
-                .await
-                .map_err(ErrorWrapper)?;
-            Ok(result)
-        })
-    }
-
-    pub fn get_energy_data<'a>(
-        &'a self,
-        py: Python<'a>,
-        interval: PyEnergyDataInterval,
-        start_date: NaiveDate,
-        end_date: Option<NaiveDate>,
-    ) -> PyResult<&'a PyAny> {
-        let handler = self.handler.clone();
-        pyo3_asyncio::tokio::future_into_py(py, async move {
-            let interval = match interval {
-                PyEnergyDataInterval::Hourly => EnergyDataInterval::Hourly {
-                    start_date,
-                    end_date: end_date.unwrap_or(start_date),
-                },
-                PyEnergyDataInterval::Daily => EnergyDataInterval::Daily { start_date },
-                PyEnergyDataInterval::Monthly => EnergyDataInterval::Monthly { start_date },
-            };
-
-            let result = handler
-                .lock()
-                .await
-                .get_energy_data(interval)
-                .await
-                .map_err(ErrorWrapper)?;
-            Ok(result)
+            Ok(())
         })
     }
 }