+++ tapo/src/responses/device_info_result/light.rs
@@ -3,8 +3,9 @@ use serde::{Deserialize, Serialize};
 use crate::error::Error;
 use crate::responses::{decode_value, DecodableResultExt, DefaultStateType, TapoResponseExt};
 
-/// Device info of Tapo L510 and L610. Superset of [`crate::responses::DeviceInfoGenericResult`].
+/// Device info of Tapo L510, L520 and L610. Superset of [`crate::responses::DeviceInfoGenericResult`].
 #[derive(Debug, Clone, Serialize, Deserialize)]
+#[cfg_attr(feature = "python", pyo3::prelude::pyclass(get_all))]
 #[allow(missing_docs)]
 pub struct DeviceInfoLightResult {
     //
@@ -50,6 +51,18 @@ pub struct DeviceInfoLightResult {
     pub default_states: DefaultLightState,
 }
 
+#[cfg(feature = "python")]
+#[pyo3::pymethods]
+impl DeviceInfoLightResult {
+    /// Gets all the properties of this result as a dictionary.
+    pub fn to_dict(&self, py: pyo3::Python) -> pyo3::PyResult<pyo3::Py<pyo3::types::PyDict>> {
+        let value = serde_json::to_value(self)
+            .map_err(|e| pyo3::exceptions::PyException::new_err(e.to_string()))?;
+
+        crate::python::serde_object_to_py_dict(py, &value)
+    }
+}
+
 impl TapoResponseExt for DeviceInfoLightResult {}
 
 impl DecodableResultExt for DeviceInfoLightResult {
@@ -63,6 +76,7 @@ impl DecodableResultExt for DeviceInfoLightResult {
 
 /// Light Default State.
 #[derive(Debug, Clone, Serialize, Deserialize)]
+#[cfg_attr(feature = "python", pyo3::prelude::pyclass(get_all))]
 #[allow(missing_docs)]
 pub struct DefaultLightState {
     pub r#type: DefaultStateType,
@@ -71,6 +85,7 @@ pub struct DefaultLightState {
 
 /// Light State.
 #[derive(Debug, Clone, Serialize, Deserialize)]
+#[cfg_attr(feature = "python", pyo3::prelude::pyclass(get_all))]
 #[allow(missing_docs)]
 pub struct LightState {
     pub brightness: u8,